requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0
numpy==1.24.3
